+++ old-ui/app/components/pending-typed-msg.js
@@ -1,7 +1,14 @@
 import React, { Component } from 'react'
 import PendingTxDetails from './pending-typed-msg-details'
+import PropTypes from 'prop-types'
 
 export default class PendingMsg extends Component {
+  static propTypes = {
+    txData: PropTypes.object,
+    cancelTypedMessage: PropTypes.function,
+    signTypedMessage: PropTypes.function,
+  }
+
   render () {
     var state = this.props
     var msgData = state.txData