+++ old-ui/app/components/send/send-contract.js
@@ -8,11 +8,10 @@ import ErrorComponent from '../error'
 import ToastComponent from '../toast'
 import Select from 'react-select'
 import actions from '../../../../ui/app/actions'
-import { AbiCoder } from 'web3-eth-abi'
+import abi from 'web3-eth-abi'
 import Web3 from 'web3'
 import copyToClipboard from 'copy-to-clipboard'
 import CopyButton from '../copy/copy-button'
-const abiEncoder = new AbiCoder()
 
 class SendTransactionField extends Component {
 	constructor (props) {
@@ -430,7 +429,7 @@ class SendTransactionScreen extends PersistentForm {
 		const inputValuesArray = Object.keys(inputValues).map(key => inputValues[key])
 		let txData
 		try {
-			txData = abiEncoder.encodeFunctionCall(methodABI, inputValuesArray)
+			txData = abi.encodeFunctionCall(methodABI, inputValuesArray)
 			this.props.hideWarning()
 		} catch (e) {
 			this.props.hideToast()