+++ test/e2e/test-cases/import-contract-account.spec.js
@@ -1,6 +1,7 @@
 const assert = require('assert')
 const clipboardy = require('clipboardy')
 const { menus, screens, elements, NETWORKS } = require('../elements')
+const { main } = screens
 let abiClipboard
 
 const importContractAccount = async (f, account1, getCreatedAccounts) => {
@@ -36,16 +37,16 @@ const importContractAccount = async (f, account1, getCreatedAccounts) => {
       })
 
       it('ABI of Proxy + Implementation is fetched and matches the pattern', async () => {
-        await f.delay(5000)
+        await f.delay(10000)
         const field = await f.waitUntilShowUp(screens.importAccounts.contractABI)
         abiClipboard = await field.getText()
-        console.log(abiClipboard)
         assert.deepEqual(JSON.parse(abiClipboard), joinedABI, "ABI isn't fetched")
       })
 
       it("Click button 'Import', main screen opens", async () => {
         const button = await f.waitUntilShowUp(screens.importAccounts.buttonImport)
         await f.click(button)
+        await f.delay(7000)
         const ident = await f.waitUntilShowUp(screens.main.identicon, 20)
         assert.notEqual(ident, false, "main screen isn't opened")
       })
@@ -93,13 +94,11 @@ const importContractAccount = async (f, account1, getCreatedAccounts) => {
         assert.equal(text1, 'PROXY', 'label incorrect')
       })
       it('Delete imported account', async () => {
-        await f.waitUntilShowUp(menus.account.delete)
-        const items = await f.driver.findElements(menus.account.delete)
-        await items[accountPosition].click()
-        const button = await f.waitUntilShowUp(screens.deleteImportedAccount.buttons.yes)
-        await button.click()
-        const buttonArrow = await f.waitUntilShowUp(screens.settings.buttons.arrow)
-        await buttonArrow.click()
+        const deleteButton = await f.waitUntilShowUp(menus.account.delete)
+        await deleteButton.click()
+        const yesButton = await f.waitUntilShowUp(screens.deleteImportedAccount.buttons.yes)
+        await yesButton.click()
+        await f.driver.findElements(main.container)
         const identicon = await f.waitUntilShowUp(screens.main.identicon)
         assert.notEqual(identicon, false, 'main screen didn\'t opened')
       })
@@ -117,15 +116,10 @@ const importContractAccount = async (f, account1, getCreatedAccounts) => {
         await f.setProvider(NETWORKS.ROPSTEN)
         const menu = await f.waitUntilShowUp(menus.account.menu)
         await menu.click()
-        const item = await f.waitUntilShowUp(menus.account.import2)
-        await item.click()
+        const importItem = await f.waitUntilShowUp(menus.account.import2)
+        await importItem.click()
         const importAccountTitle = await f.waitUntilShowUp(screens.importAccounts.title)
         assert.equal(await importAccountTitle.getText(), screens.importAccounts.textTitle)
-      })
-
-      it("Warning's  text is correct", async () => {
-        const field = await f.waitUntilShowUp(screens.importAccounts.warning)
-        assert.equal(await field.getText(), 'Imported accounts will not be associated with your originally created Nifty Wallet account seedphrase.', "incorrect warning's text")
       })
 
       it("Select type 'Contract'", async () => {
@@ -208,6 +202,7 @@ const importContractAccount = async (f, account1, getCreatedAccounts) => {
       it("Click button 'Import', main screen opens", async () => {
         const button = await f.waitUntilShowUp(screens.importAccounts.buttonImport)
         await f.click(button)
+        await f.delay(5000)
         const ident = await f.waitUntilShowUp(screens.main.identicon, 20)
         assert.notEqual(ident, false, "main screen isn't opened")
       })
@@ -865,8 +860,7 @@ const importContractAccount = async (f, account1, getCreatedAccounts) => {
         await items[1].click()
         const button = await f.waitUntilShowUp(screens.deleteImportedAccount.buttons.yes)
         await button.click()
-        const buttonArrow = await f.waitUntilShowUp(screens.settings.buttons.arrow)
-        await buttonArrow.click()
+        await f.driver.findElements(main.container)
         const identicon = await f.waitUntilShowUp(screens.main.identicon)
         assert.notEqual(identicon, false, 'main screen didn\'t opened')
       })